### 2024-03-13 #6476 @editorialbot
[PRE REVIEW]: DeepBench: A simulation package for physical benchmarking data
**Submitting author:** @voetberg (<a href="http://orcid.org/0009-0005-2715-4709">M. Voetberg</a>)
**Repository:** https://github.com/deepskies/DeepBench/
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.2
