### 2024-05-23 #6476 @plaplant
As an editorial note, @jwuphysics has been affiliated with DeepSkies in the past. However, he has not contributed directly to this submission, and he believes that he can review the software fairly. As such, I do not believe there is a conflict of interest in this review.
