### 2024-05-24 #6774 @plaplant
@jwuphysics @apoudel2014 thanks for agreeing to review! When you get a chance, please respond to this issue with the comment:
This will make your review checklist, after which you can begin going through each of the items. If you see things that should be addressed, please feel free to open an issue on [the software repository](https://github.com/deepskies/DeepBench) that references this review issue. Please let me know if you have any questions!
