+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "DeepBench"
-version = "0.2.2"
+version = "0.2.3"
 description = "Physics Benchmark Dataset Generator"
 license = "Apache 2.0"
 authors = ["M. Voetberg <maggiev@fnal.gov>", "Ashia Livaudais", "Becky Nevin", "Omari Paul", "Brian Nord"]
