+++ pyproject.toml
@@ -11,10 +11,10 @@ repository = "https://github.com/deepskies/DeepBench"
 documentation = "https://deepbench.readthedocs.io/en/latest/"
 
 [tool.poetry.dependencies]
-python = ">=3.8,<4.0,"
-numpy = "^1.24.3"
+python = ">=3.10,<4.0,"
+numpy = "^1.25.0"
 matplotlib = "^3.7.1"
-scikit-image = "^0.20.0"
+scikit-image = "^0.23.0"
 astropy = "^5.2.2"
 autograd = "^1.5"
 pyyaml = "^6.0"
