+++ .readthedocs.yaml
@@ -8,10 +8,8 @@ build:
     python: "3.10"
   jobs:
     post_create_environment:
-      - pip install poetry
-      - poetry config virtualenvs.create false
+      - pip install poetry 
     post_install:
-      - poetry install
-
+      - VIRTUAL_ENV=$READTHEDOCS_VIRTUALENV_PATH poetry install --with dev
 sphinx:
   configuration: docs/conf.py
