### 2024-12-23 #6774 @jwuphysics
@plaplant I'm sorry for the delay, and my sincerest apologies to the authors for my tardy review. (I was on parental leave for 3 months, and lost track of my task lists...)

I re-installed and reviewed the code, reviewed the paper, and found it all to be of excellent quality. As such I can attest that all items on the checklist are good to go.

The one minor question I have--probably for the editors here--is regarding citations. It doesn't seem that the draft paper PDF has properly rendered the citations or the references section. Is this just a pandoc issue? The citations in the text do seem to conform to the requested style.

Best,
John
