### 2024-12-25 #6774 @apoudel2014
I apologize for the significant delay in reviewing this work.

I have carefully reviewed the code and the paper, and I found both to be of excellent quality. Based on my review, I can confirm that all items on the checklist are satisfactory.

I have a few minor recommendations regarding the presentation of the example output figures:
Figure 1: Please provide a more descriptive caption for Figure 1. Specifically, clarify that the "single object" and "two objects" examples include both noiseless and noisy versions.
Figure 2: Please include units for the x-position on the axis. For consistency, I recommend using the term "noiseless" throughout the figure description and legend, as this aligns with the terminology used in Figure 1. This will simplify any necessary changes to the figure.

Thanks,
Amit
