### 2024-12-28 #6774 @plaplant
@jwuphysics @apoudel2014 thank you very much for your reviews! We appreciate the time and effort required to review, and are grateful for your contribution to this submission.

@voetberg thank you for your patience with the review process. The two reviewers have signed off with minor suggestions, so assuming those will be addressed, we can begin the final acceptance checks before publication. In a separate comment, I will make a checklist of issues that need to be addressed, and will include the issues pointed out by the reviewers. Let me know if you have any questions!
