### 2025-01-03 deepskies/DeepBench#147 @voetberg
Improve figures (description and axis labels)
As per the comments on the joss review:
https://github.com/openjournals/joss-reviews/issues/6774#issuecomment-2564386938
