### 2025-01-03 22c06a1
Review comments on figure captions
