+++ paper/paper.md
@@ -10,11 +10,11 @@ authors:
   - name: M. Voetberg
     orcid: 0009-0005-2715-4709
     equal-contrib: true
-    affiliation: "1" 
-  - name: Ashia Livaudais 
+    affiliation: "1"
+  - name: Ashia Livaudais
     orcid: 0000-0003-3734-335X
     equal-contrib: true
-    affiliation: "1" 
+    affiliation: "1"
   - name: Becky Nevin
     orcid: 0000-0003-1056-8401
     equal-contrib: false
@@ -49,7 +49,7 @@ There are multiple open problems and issues that are critical for the machine le
 
 The physical sciences community lacks sufficient datasets and software as benchmarks for the development of statistical and machine learning models.  In particular, there currently does not exist simulation software that generates data underpinned by physical principles and that satisfies the following criteria:
 
-* multi-domain 
+* multi-domain
 * multi-purpose
 * fast
 * reproducible
@@ -60,13 +60,13 @@ The physical sciences community lacks sufficient datasets and software as benchm
 
 ## Related Work
 
-There are many benchmarks -- datasets and simulation software -- widely used for model building in machine learning, statistics, and the physical sciences. 
+There are many benchmarks -- datasets and simulation software -- widely used for model building in machine learning, statistics, and the physical sciences.
 First, benchmark datasets of natural images include MNIST `[@dengMnistDatabaseHandwritten2012c]`, CIFAR `[@krizhevskyCIFAR10CanadianInstitute2017a]`, Imagenet `[@russakovskyImageNetLargeScale2015a]`. Second, there are several large astronomical observation datasets -- CfA Redshift Survey `[@huchraSurveyGalaxyRedshifts1983]`, Sloan Digital Sky Survey `[@yorkSloanDigitalSky2000]`, and Dark Energy Survey `[@abbottDARKENERGYSURVEY]`. Third, many n-body cosmology simulation data sets serve as benchmarks -- e.g., Millennium `[@springelCosmologicalSimulationCode2005]`, Illustris `[@vogelsbergerIntroducingIllustrisProject2014b]`, EAGLE `[@schayeEAGLEProjectSimulating2015]`, Coyote `[@heitmannCoyoteUniversePrecision2010]`, Bolshoi `[@klypinDARKMATTERHALOS2011]`, CAMELS `[@villaescusa-navarroCAMELSProjectCosmology2021]`, Quijote `[@villaescusa-navarroQuijoteSimulations2020]`. Fourth, there have been multiple astronomy data set challenges that can be considered benchmarks for analysis comparison -- e.g., PLAsTiCC `[@hlozekResultsPhotometricLSST2020a]`, The Great08 Challenge `[@bridleHandbookGREAT08Challenge2009a]`, and the Strong Gravitational Lens Challenge `[@metcalfStrongGravitationalLens2019c]`.  Fifth, there are multiple software that generate simulated data for astronomy and cosmology -- e.g., astropy `[@theastropycollaborationAstropyCommunityPython2013a]`, galsim `[@roweGalSimModularGalaxy2015]`, lenstronomy `[@birrerLenstronomyMultipurposeGravitational2018a]`, deeplenstronomy `[@morganDeeplenstronomyDatasetSimulation2021a]`, CAMB `[@CAMBInfo, @lewisEfficientComputationCMB2000]`, Pixell `[@WelcomePixellDocumentation]`, SOXs `[@SOXSSimulatedObservations]`. Finally, particle physics projects use standard codebases for simulations -- e.g., GEANT `[@piaGeant4ScientificLiterature2009]`, GENIE `[@andreopoulosGENIENeutrinoMonte2015]`, and PYTHIA `[@sjostrandPYTHIAEventGenerator20
[…]
