+++ paper/figures/code/example_pendulums.py
@@ -26,7 +26,9 @@ pendulum = Pendulum(
 # Plot that against the time and with scatter and line options
 pendulum_noiseless = pendulum.create_object(time, noiseless=True)
 subplots[0].plot(time, pendulum_noiseless, color="black")
-subplots[0].scatter(time, pendulum_noiseless, color="black", label="Noiseless")
+subplots[0].scatter(
+    time, pendulum_noiseless, color="black", label="Noiseless", marker=">"
+)
 
 # Use the noiseless=False to do the same with a noiseless pendulum
 pendulum_noisy = pendulum.create_object(time, noiseless=False)
@@ -46,22 +48,22 @@ pendulum = HamiltonianPendulum(
     },
 )
 
-# Cacluate the pendulum positions and engeries
+# Calculate the pendulum positions and energies
 pendulum_data = pendulum.create_object(time)
 
 # Plot the line and scatterplot versions of the position wrt time
 subplots[1].plot(pendulum_data[4], pendulum_data[0], color="black")
 subplots[1].scatter(
-    pendulum_data[4], pendulum_data[0], color="black", label="Noiseless"
+    pendulum_data[4], pendulum_data[0], color="black", label="Noiseless", marker=">"
 )
 
-# Repeat the process with the noisely pendulum
+# Repeat the process with the noisy pendulum
 pendulum = HamiltonianPendulum(
     pendulum_arm_length=10.0,
     starting_angle_radians=np.pi / 4,
     acceleration_due_to_gravity=9.8,
     noise_std_percent={
-        "pendulum_arm_length": 0.2,
+        "pendulum_arm_length": 0.1,
         "starting_angle_radians": 0.0,
         "acceleration_due_to_gravity": 0.0,
     },
@@ -81,9 +83,9 @@ for plot in subplots.ravel():
     # plot.set(xticks=[], yticks=[])
 
     plot.set_xlabel("Time (s)")
-    plot.set_ylabel("X Position")
+    plot.set_ylabel("X Position (m)")
 
 # Assign legend location
 subplots[1].legend(loc="center left", bbox_to_anchor=(1.02, 1))
 
-plt.savefig("../pendulums.png")
+plt.savefig("./pendulums.png")
