### 2025-01-03 #6774 @voetberg
@plaplant Thank you for this checklist! I have put in a pull request for the caption/figure changes and will archive a release once that is merged and my collaborators agree on the changes.

Thank you very much to @jwuphysics and @apoudel2014 for your reviews!

I will second @jwuphysics ' question
> The one minor question I have--probably for the editors here--is regarding citations. It doesn't seem that the draft paper PDF has properly rendered the citations or the references section. Is this just a pandoc issue? The citations in the text do seem to conform to the requested style.

I wasn't able to figure out if the citation rendering incorrectly in the article proof is the fault of the hook here or something I need to fix
