### 2025-01-25 bcb60f7
Fix citation issues and make some proofing light edits
