### 2025-01-25 #6774 @plaplant
@voetberg apologies for the delay, as I was only recently able to dig in and figure out why the references weren't rendering correctly. I've made a [small PR](https://github.com/deepskies/DeepBench/pull/149) that should fix the citation rendering issue (and a few other small manuscript issues). When you get a chance, please review the PR and merge it so we can make sure that fixes things.

Related to a few of the issues on my checklist above, please also reply to this thread with:
- a new release of the software incorporating the changes recommended by the reviewers.
- the version number of this new release.
- a link to an archived version of the new release of the repository (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or another suitable long-term hosting platform). Note that the author information of the archived version must match that of the manuscript.

Thanks! Let me know if you have any questions.
