### 2025-02-10 #6774 @voetberg
@plaplant Thanks for all the assistance with the review! I completed the items on the checklist.

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

[Release 1.0-joss - view the release here](https://github.com/deepskies/DeepBench/releases/tag/v1.0-joss)

> Archive the release on Zenodo/figshare/etc and post the DOI here.

[10.5281/zenodo.14845251](https://zenodo.org/records/14845251)
