+++ README.md
@@ -1,7 +1,10 @@
 ![GitHub Workflow Status](https://github.com/deepskies/DeepBench/actions/workflows/test-bench.yml/badge.svg?label=test)
 [![License](https://img.shields.io/badge/License-Apache_2.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
- [![PyPI version](https://badge.fury.io/py/deepbench.svg)](https://badge.fury.io/py/deepbench)
+[![PyPI version](https://badge.fury.io/py/deepbench.svg)](https://badge.fury.io/py/deepbench)
 [![Documentation Status](https://readthedocs.org/projects/deepbench/badge/?version=latest)](https://deepbench.readthedocs.io/en/latest/?badge=latest)
+[![status](https://joss.theoj.org/papers/300762982613649881f8b6a08dabd33e/status.svg)](https://joss.theoj.org/papers/300762982613649881f8b6a08dabd33e)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14845251.svg)](https://doi.org/10.5281/zenodo.14845251)
+
 
 ### What is it?
 Simulation library for very simple simulations to *benchmark* machine learning algorithms.
